+++ src/dialogs/joplinimportdialog.h
@@ -2,6 +2,8 @@
 
 #include "masterdialog.h"
 
+#include <entities/notesubfolder.h>
+
 namespace Ui {
 class JoplinImportDialog;
 }
@@ -24,8 +26,7 @@ class JoplinImportDialog : public MasterDialog {
     ~JoplinImportDialog() override;
 
     int getImportCount() const;
-
-   private slots:
+private slots:
     void on_directoryButton_clicked();
 
     void on_importButton_clicked();
@@ -33,15 +34,20 @@ class JoplinImportDialog : public MasterDialog {
    private:
     Ui::JoplinImportDialog *ui;
     int _importCount;
+    QString _dirPath;
     QHash <QString, QString> _noteData;
+    QHash <QString, QString> _folderData;
     QHash <QString, QString> _tagData;
     QHash <QString, QStringList> _tagAssignmentData;
     QHash <QString, QString> _imageData;
     QHash <QString, QString> _attachmentData;
+    QHash <QString, NoteSubFolder> _importedFolders;
 
     bool importNote(const QString& id, const QString& text,
                     const QString& dirPath);
     void tagNote(const QString& id, const Note& note);
     void handleImages(Note& note, const QString& dirPath);
     void handleAttachments(Note& note, const QString& dirPath);
+    bool importFolders();
+    NoteSubFolder importFolder(const QString &id, const QString &text);
 };